pyproject.toml
[project]
name = "retrieval-process-docs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "celery>=5.6.3",
    "docx2txt>=0.9",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.26",
    "langchain-openai>=1.1.12",
    "langchain-pinecone>=0.2.13",
    "langchain-text-splitters>=1.1.1",
    "llama-index-core>=0.14.20",
    "llama-index-embeddings-openai>=0.6.0",
    "llama-index-llms-openai>=0.7.5",
    "llama-index-readers-file>=0.6.0",
    "llama-index-vector-stores-pinecone>=0.8.0",
    "openai>=2.30.0",
    "opentelemetry-api>=1.40.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.40.0",
    "opentelemetry-sdk>=1.40.0",
    "pinecone>=7.3.0",
    "prometheus-fastapi-instrumentator>=7.1.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pypdf>=6.9.2",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-env>=1.6.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "redis>=7.4.0",
    "structlog>=25.5.0",
    "tenacity>=9.1.4",
    "tiktoken>=0.12.0",
    "uvicorn>=0.44.0",
]
